productimporter/requirements.txt
amqp==2.4.2
aniso8601==3.0.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-admin==1.3.2
django-background-tasks==1.2.0
django-compat==1.0.15
django-cors-headers==2.5.2
django-excel-response2==2.0.8
django-filter==2.1.0
django-graphql-jwt==0.2.1
django-six==1.0.4
djangorestframework==3.9.2
dnspython==1.16.0
eventlet==0.24.1
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
kombu==4.5.0
monotonic==1.5
numpy==1.16.2
pandas==0.24.2
promise==2.2.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
Rx==1.6.1
screen==1.0.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
vine==1.3.0
waitress==1.2.1
xlwt==1.3.0
